#258 rename modal
